backend_flask/requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
duckduckgo_search
requests 